Add files via upload
Took some liberties with E-F sub styling for episode 1. For example during the memory sequence of Sai and Hikaru (starting around 4:36 on E-F), E-F did this with Default styling but it really should be Background styling. 

They even apply the background styling sometimes (e.g. "What are you?" at 4:41) so I guess they just forgot to apply it in places. They also applied Background to Sai talking (4:37) instead of the Sai styling which makes more sense.

Around 11:54 they styled Hikaru as Sai so that got changed to Background which again seems to make the most sense. I started to apply the Thinking style starting with Touya saying "That's the way a beginner holds a stone...". E-F only started doing the Thinking style in later episodes but no reason to do it from the start now. 

As things progressed I realise I should be doing small improvements like this.